src/phase2/backend/requirements.txt
azure-ai-projects
azure-ai-inference[prompts]
azure-core
azure-identity
azure-monitor-opentelemetry
azure-search-documents
pandas
python-dotenv
opentelemetry-api

flask
flask-cors